+++ src/lib.rs
@@ -102,6 +102,7 @@ pub use crate::algorithms::connectivity::reachability_matrix::reachability_matri
 pub use crate::algorithms::connectivity::strong::strongly_connected_components;
 pub use crate::algorithms::connectivity::transitive_closure::transitive_closure;
 pub use crate::algorithms::games::barabasi::barabasi_game_bag;
+pub use crate::algorithms::games::callaway_traits::callaway_traits_game;
 pub use crate::algorithms::games::chung_lu::{ChungLuVariant, chung_lu_game};
 pub use crate::algorithms::games::erdos_renyi::{erdos_renyi_gnm, erdos_renyi_gnp};
 pub use crate::algorithms::games::establishment::establishment_game;